pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oceanpath"
version = "0.1.0"
description = "MIL pretraining, benchmarking, and deployment for computational pathology"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "h5py>=3.15.1",
    "huggingface-hub>=1.4.1",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "lightning>=2.6.1",
    "litlogger>=0.1.7",
    "numpy>=2.2.6",
    "nystrom-attention>=0.0.14",
    "omegaconf>=2.3.0",
    "onnx>=1.20.1",
    "onnxruntime<1.24",
    "opensdpc",
    "pandas>=2.3.3",
    "pyarrow>=23.0.1",
    "scikit-learn>=1.7.2",
    "torch>=2.10.0",
    "torchmetrics>=1.8.2",
    "tqdm>=4.67.3",
    "trident",
    "umap-learn>=0.5.11",
    "uvicorn>=0.40.0",
    "wandb>=0.25.0",
]

[project.optional-dependencies]
serve = [
    "onnxruntime-gpu<1.24",
    "uvicorn[standard]>=0.40.0",
]

[tool.uv.sources]
opensdpc = { git = "https://github.com/WonderLandxD/opensdpc" }
trident = { git = "https://github.com/mahmoodlab/TRIDENT" }

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/oceanpath"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes (unused imports, undefined names)
    "I",     # isort (import sorting)
    "UP",    # pyupgrade (Python 3.10+ idioms)
    "B",     # flake8-bugbear (common bugs)
    "SIM",   # flake8-simplify (simplifiable code)
    "RUF",   # ruff-specific rules
    "C4",    # flake8-comprehensions (better list/dict/set comprehensions)
    "PIE",   # flake8-pie (unnecessary code)
    "PT",    # flake8-pytest-style (pytest best practices)
    "RET",   # flake8-return (return statement issues)
    "TCH",   # flake8-type-checking (TYPE_CHECKING block optimization)
    "NPY",   # numpy-specific rules
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "B008",    # do not perform function calls in argument defaults (Hydra uses this)
    "B905",    # zip without strict= (Python 3.10 compat)
    "SIM108",  # use ternary operator (readability preference)
    "RET504",  # unnecessary assignment before return (readability)
    "UP007",   # use X | Y for union types (keep Optional for clarity)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "B011",   # assert False is fine in tests
    "PT011",  # pytest.raises match is optional in tests
    "SIM300", # yoda conditions fine in assertions
]
"scripts/*" = [
    "B008",   # Hydra uses function defaults
]

[tool.ruff.lint.isort]
known-first-party = ["oceanpath"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ═════════════════════════════════════════════════════════════════════════════
# Mypy — static type checking
# ═════════════════════════════════════════════════════════════════════════════

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false       # too strict for research code initially
check_untyped_defs = true           # but DO check bodies of untyped functions
ignore_missing_imports = true        # third-party stubs often missing
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "oceanpath.models.*"
disallow_untyped_defs = true         # strict typing for model interfaces

[[tool.mypy.overrides]]
module = "oceanpath.eval.*"
disallow_untyped_defs = true         # strict typing for evaluation

# ═════════════════════════════════════════════════════════════════════════════
# Pytest
# ═════════════════════════════════════════════════════════════════════════════

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short -x --strict-markers"
markers = [
    "slow: marks tests that take >10s (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require CUDA",
    "integration: marks tests that need external resources (memmap, models)",
]
filterwarnings = [
    "ignore::UserWarning:umap",
    "ignore::DeprecationWarning",
]

# ═════════════════════════════════════════════════════════════════════════════
# Coverage
# ═════════════════════════════════════════════════════════════════════════════

[tool.coverage.run]
source = ["src/oceanpath"]
omit = [
    "*/serving/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ ==",
    "raise NotImplementedError",
]
show_missing = true
skip_empty = true

[tool.setuptools.packages.find]
where = ["src"]
